Redo!
Apparently we were using the wrong FTC repository for this year, ftc_app. The current one is skystone. I backed up the files that we had, and completely wiped out the ftc_app fork. with this new repo we will be able to use CV (vuforia, dogecv), I also reversed all motors in encoder_test because everything the bot was doing was backwards